Convert all decimal data types into floats

+++ tracker/models.py
@@ -8,8 +8,8 @@ class Crypto(models.Model):
 
 class Operation(models.Model):
     symbol = models.ForeignKey(Crypto, on_delete=models.CASCADE)
-    cryptoQuantity = models.DecimalField(max_digits=20, decimal_places=4)
-    price = models.DecimalField(max_digits=20, decimal_places=8)
+    cryptoQuantity = models.FloatField()
+    price = models.FloatField()
     isSell = models.BooleanField()
     def __str__(self):
         symbol = '-' if self.isSell else '+'
